+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Refract Python Changelog
 
-## Master
+## 0.3.1 (2017-06-02)
 
 ### Enhancements